# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Admin panel API for a Dune Awakening private server.",
        "title": "dune-admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/update/apply": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Download and apply the latest release, then restart",
                "parameters": [
                    {
                        "description": "Set force=true to reinstall even when up to date",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.updateApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.updateApplyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/update/check": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Check for a newer release on GitHub",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.updateCheckResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.updateApplyRequest": {
            "type": "object",
            "properties": {
                "force": {
                    "type": "boolean"
                }
            }
        },
        "main.updateApplyResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "updated": {
                    "type": "boolean"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "main.updateCheckResponse": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "string"
                },
                "latest": {
                    "type": "string"
                },
                "needs_update": {
                    "type": "boolean"
                },
                "release_url": {
                    "type": "string"
                }
            }
        },
...
    }
}